Refactored services in paperlessREST in order to use interfaces when possible
