+++ paperlessREST/src/main/java/at/fhtw/swen3/paperless/services/messageQueue/MessageQueueService.java
@@ -1,0 +1,7 @@
+package at.fhtw.swen3.paperless.services.messageQueue;
+
+public interface MessageQueueService {
+
+    public void processMessage(String message);
+
+}
