+++ paperlessREST/src/main/java/at/fhtw/swen3/paperless/services/minio/DocumentStoreService.java
@@ -1,0 +1,9 @@
+package at.fhtw.swen3.paperless.services.minio;
+
+import org.springframework.web.multipart.MultipartFile;
+
+public interface DocumentStoreService {
+
+    public void handleFileUpload(MultipartFile document);
+
+}
