+++ paperlessREST/src/test/java/at/fhtw/swen3/paperless/services/DocumentMapperTests.java
@@ -8,6 +8,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+@SpringBootTest
 public class DocumentMapperTests {
 
     @Test
